requirements.txt
aiogram<4.0
pydantic-settings

python-dotenv~=1.0.1
aiosqlite~=0.20.0